Create a new docker image which does not have the redundent scalapack installation.
Version 2 has qt5 installed. Noticed that this is not the qt5 version compatible with Elemental. Elemental can use QT5.1.1. The existing QT5 is some other version.

There is no guarantee that in the future I'll add the compatible version since compiling it has been very difficult and I failed in installing from the .run file
due to the fact that it's very difficult to start a X server inside the docker.

I have no idea why I in the end added the qt5 to this image. I just feel that perhaps it might be of some help since we might in the end have to write some C++ code
for this project.

Notice that even though Elemental can not use this version of QT. QT can surely use this version of Elemental so long as we convert the Elemental data to a format
that is compatible with this version of QT.